# PERF: speed up nansum for integer inputs with axis != None by up to ~6x